server now differentiates between ACK type packets (unreliable) and DATA type packets (reliable). lua scripts are working in bizhawk. TODO: add client-side packet re-sends if no ACK has arrived for them. TODO: make sure lib can return the session key that it joined and have a reset function to prepare for the next match

+++ src/threads/thread_message.rs
@@ -4,7 +4,7 @@ pub enum ThreadMessage {
     Tick(Box<dyn FnOnce() + Send>),
     ClientPacket {
         socket_address: std::net::SocketAddr,
-        id: u64,
+        id: u32,
         packet: ClientPacket
     }
 }
